refactor PrefillRunner
refactor PrefillRunner for qwen2 and llama
